fix: use a big-int to match key_id primary key

+++ database/migrations/2019_03_35_000002_update_logger_tables.php
@@ -16,7 +16,7 @@ class UpdateLoggerTables extends Migration
 
         Schema::table('audit_changes', function (Blueprint $table) {
             $table->boolean('processed')->default(false);
-            $table->integer('key_id', false, true);
+            $table->bigInteger('key_id', false, true);
 
             $table
                 ->foreign('key_id')
